% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustohlc.R
\name{adjustohlc}
\alias{adjustohlc}
\title{Return a dataframe with adjusted OHLC for given tickers}
\usage{
adjustohlc(tickers, start, end)
}
\arguments{
\item{tickers}{A ticker in quotes: "AAPL", or subset of tickers: c("AAPL", "TSLA") for example}

\item{start}{Start date}

\item{end}{End date description}
}
\value{
An table with adjusted OHLC prices
}
\description{
This function takes 3 inputs: a ticker or list of tickers, a start date and an end date and returns the adjusted OHLC
table.
}
\examples{
adjustohlc("AAPL", "2020-01-01", "2023-02-01")

}
