% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morningstar.R
\name{getAgPrice}
\alias{getAgPrice}
\title{Morning Star Commodities API call}
\usage{
getAgPrice(
  feed = "MGEX_Futures_continuous",
  contract = "MWE_001_Month",
  from = "2000-01-01",
  iuser = "a@abc.com",
  ipassword = "password"
)
}
\arguments{
\item{feed}{Motningstar Feed Table. \code{character}}

\item{contract}{Morningstar key. \code{character}}

\item{from}{From date yyyy-mm-dd. \code{character}}

\item{iuser}{Morningstar username/email as character - locally sourced for examples. \code{character}}

\item{ipassword}{Morningstar password as character - locally sourced for examples. \code{character}}
}
\value{
wide data frame \code{tibble}
}
\description{
Returns data from Morningstar API.
Credentials are needed for Morningstar.
}
\section{Current feeds supported}{

\itemize{
\item MGEX_Futures_continuous and MGEX_Futures
\item ICE_Canada_FuturesPrices_continuous
}
}

\examples{
\dontrun{

getAgPrice(feed = "MGEX_Futures_continuous",
          contract = "MWE_001_Month",
          from = "2000-01-01",
          iuser = username,
          ipassword = password)

getAgPrice(feed = "MGEX_Futures",
          contract = "@AJC6F",
          from = "2000-01-01",
          iuser = username,
          ipassword = password)

getAgPrice(feed = "ICE_Canada_FuturesPrices_continuous",
           contract = "BW_001_Month",
           from = "2000-01-01",
           iuser = username,
           ipassword = password)
}
}
\author{
Brooklyn Holt, developed from Philippe Cote's getPrice
}
