% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morningstar.R
\name{getAgPrices}
\alias{getAgPrices}
\title{Morning Star Commodities API multiple calls}
\usage{
getAgPrices(
  feed = "ICE_Canada_FuturesPrices_continuous",
  contracts = c("BW_001_Month", "RS_001_Month"),
  from = "2000-01-01",
  iuser = "a@abc.com",
  ipassword = "password"
)
}
\arguments{
\item{feed}{Motningstar Feed Table. \code{character}}

\item{contracts}{Symbols vector. \code{character}}

\item{from}{From date yyyy-mm-dd. \code{character}}

\item{iuser}{Morningstar username/email as character - locally sourced for examples. \code{character}}

\item{ipassword}{Morningstar password as character - locally sourced for examples. \code{character}}
}
\value{
wide data frame \code{tibble}
}
\description{
Returns multiple Morningstar calls from the getAgPrice functions.
Credentials are needed for Morningstar.
}
\section{Current feeds supported}{

\itemize{
\item MGEX_Futures_continuous and MGEX_Futures
\item ICE_Canada_FuturesPrices_continuous
}
}

\examples{
\dontrun{
getAgPrices(feed = "ICE_Canada_FuturesPrices_continuous",
            contracts = c("BW_001_Month", "RS_001_Month"),
            from = "2000-01-01",
            iuser = username,
            ipassword = password)
}

}
\author{
Brooklyn Holt, developed from Philippe Cote's getPrices
}
