% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollReg.R
\name{rollReg}
\alias{rollReg}
\title{Univariant regression rolling regression}
\usage{
rollReg(data, window, x, y)
}
\arguments{
\item{data}{A wide data frame with columns for dependent and independent variables.}

\item{window}{The size of the rolling window.}

\item{x}{The name of the independent variable.}

\item{y}{The name of the dependent variable.}
}
\value{
The input data frame with additional columns for rolling regression coefficients and P-Values.

Appends columns with rolling regression coefficients and P-Values
}
\description{
This function takes a wide data frame and computes regression coefficients
and P-Values on a rolling basis on the defined dependent and independent columns
}
\examples{
rollReg(mtcars, 5, "hp", "mpg")


}
