% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide2gg.R
\name{wide2gg}
\alias{wide2gg}
\title{Create a chart using ggplot2 from a wide dataframe}
\usage{
wide2gg(data)
}
\arguments{
\item{data}{A wide dataframe with columns 'date', series of names}
}
\value{
A ggplot chart
}
\description{
This function takes a wide dataframe as input and generates a ggplot line plot
using the ggplot2 library.
}
\examples{
wide2gg(stocks)


}
